+++ mdsf/src/tools/duster_fix.rs
@@ -1,0 +1,23 @@
+///
+/// THIS FILE IS GENERATED USING CODE - DO NOT EDIT MANUALLY
+///
+use crate::runners::CommandType;
+
+#[inline]
+pub fn set_args(
+    mut cmd: std::process::Command,
+    file_path: &std::path::Path,
+) -> std::process::Command {
+    cmd.arg("fix");
+    cmd.arg("--quiet");
+    cmd.arg("--no-interaction");
+    cmd.arg(file_path);
+    cmd
+}
+
+pub const COMMANDS: [CommandType; 2] = [
+    CommandType::PhpVendor("duster"),
+    CommandType::Direct("duster"),
+];
+
+pub const IS_STDIN: bool = false;